Enable multiprocessing on all platforms
Python instantiates processes differently across platforms.
On Linux, processes are instantiated using "fork",
whereas on Windows and macOS, processes are instantiated using "spawn".
Prior to this commit, multiprocessing failed on Windows and macOS,
due to the assumption that processes are always forked
(and hence inherit resources from the parent process).
The failure was related to the dynamic import of the configuration modules.
See https://book-of-gehn.github.io/articles/2022/03/06/Multiprocessing-Spawn-of-Dynamically-Imported-Code.html.
This commit re-writes multiprocessing such that processes are spawned on all platforms,
which will be the default as of Python 3.14 anyway,
see https://docs.python.org/3/library/multiprocessing.html#contexts-and-start-methods.